Web design!
Here are the final images for the home page (6 png files + 6 ai files)
You will also find the Layout for the homepage and the other pages (2
jpg files + 2 indd files). I took the 1st one, “What’s on”, but
actually there is no “news” in the content they provided us so it was
kinda hard to find what to say.
Don’t worry about the content for all the pages, I will read through
the whole ZIP file and choose what is relevant (same for the image
gallery).
Here are the fonts I used:
- Century Gothic (regular) for the navigation bar and the footer
- Tahoma (bold) for the “Book tickets”/“Donate” buttons
- Minion Pro (italic, bold, regular…) for all the rest