Added in Images and tinkered
Added in majority of images apart from logo for homepage. Also changed
widths to be less moveable depending on browser size to not stuff up
images, will need to fix responsive design issue by making them drop
down on to the next line when page gets to small. Also tried to make
rollover for buttons appear as in Mathilde's designs but couldn't make
it work so its commented for further experimentation.